feat : dataSource 검색어 삭제 함수 구현

+++ data/src/main/java/com/example/data/source/local/dao/SearchWordDao.kt
@@ -22,7 +22,7 @@ interface SearchWordDao {
     suspend fun deleteAllSearchWords()
 
     @Query("DELETE FROM SearchWordEntity WHERE id = :id")
-    suspend fun deleteSearchWordsById(id: Int)
+    suspend fun deleteSearchWordsById(id: Long)
 
     @Query("SELECT * FROM SearchWordEntity WHERE keyword = :keyword")
     suspend fun getSearchWordByKeyword(keyword: String): SearchWordEntity?
